+++ misc/dbt-materialize/dbt/adapters/materialize/impl.py
@@ -15,81 +15,16 @@
 # limitations under the License.
 
 from dbt.adapters.materialize import MaterializeConnectionManager
-from dbt.adapters.materialize.relation import (
-    MaterializeRelation,
-    MaterializeRelationType,
-)
-from dbt.adapters.postgres import PostgresAdapter, PostgresColumn
-
-MATERIALIZE_GET_COLUMNS_MACRO_NAME = "materialize_get_columns"
-MATERIALIZE_CONVERT_COLUMNS_MACRO_NAME = "sql_convert_columns_in_relation"
-MATERIALIZE_GET_SCHEMAS_MACRO_NAME = "materialize_get_schemas"
-MATERIALIZE_GET_FULL_VIEWS_MACRO_NAME = "materialize_get_full_views"
-MATERIALIZE_SHOW_VIEW_MACRO_NAME = "materialize_show_view"
+from dbt.adapters.postgres import PostgresAdapter
 
 
 class MaterializeAdapter(PostgresAdapter):
     ConnectionManager = MaterializeConnectionManager
-    Column = PostgresColumn
-    Relation = MaterializeRelation
 
-    @classmethod
-    def date_function(cls):
-        return "now()"
-
-    @classmethod
-    def is_cancelable(cls):
-        return False
-
-    def get_columns_in_relation(self, relation):
-        columns = self.execute_macro(
-            MATERIALIZE_GET_COLUMNS_MACRO_NAME, kwargs={"relation": relation}
-        )
-
-        table = []
-        for _field, _nullable, _type in columns:
-            table.append((_field, _type))
-
-        return self.execute_macro(
-            MATERIALIZE_CONVERT_COLUMNS_MACRO_NAME, kwargs={"table": table}
-        )
-
-    def list_relations_without_caching(self, schema_relation):
-        # Materialize errors if you try to list views from a schema that
-        # doesn't exist. Check that the schema exists first, returning an
-        # empty list of relations if not.
-        schemas = self.execute_macro(MATERIALIZE_GET_SCHEMAS_MACRO_NAME)
-        if schema_relation.schema not in schemas:
-            return []
-
-        full_views = self.execute_macro(
-            MATERIALIZE_GET_FULL_VIEWS_MACRO_NAME,
-            kwargs={"schema": schema_relation.schema},
-        )
-
-        relations = []
-        quote_policy = {"database": True, "schema": True, "identifier": True}
-        for view, type, materialized in full_views.rows:
-            dbt_type = "table" if materialized else "view"
-            relations.append(
-                self.Relation.create(
-                    database=schema_relation.database,
-                    schema=schema_relation.schema,
-                    identifier=view,
-                    quote_policy=quote_policy,
-                    type=dbt_type,
-                )
-            )
-        return relations
-
-    def check_schema_exists(self, database, schema):
-        return schema in self.list_schemas(database)
-
-    # jwills hacking to get stuff to work
     def _link_cached_relations(self, manifest):
-        schemas = set()
-        # only link executable nodes
-        relations_schemas = self._get_cache_schemas(manifest)
-        for relation in relations_schemas:
-            self.verify_database(relation.database)
-            schemas.add(relation.schema.lower())
+        # NOTE(benesch): this *should* reimplement the parent class's method
+        # for Materialize, but presently none of our tests care if we just
+        # disable this method instead.
+        #
+        # [0]: https://github.com/dbt-labs/dbt-core/blob/13b18654f03d92eab3f5a9113e526a2a844f145d/plugins/postgres/dbt/adapters/postgres/impl.py#L126-L133
+        pass